Add exclusion for *Html methods in MEQP1 XSS sniff
The sniff already had support for this, but it appears to have only been
used in the Magento 2 version, even though it's relevant for Magento 1
as well. Specifically, we were most often running into this when using
getChildHtml calls, which obviously shouldn't be escaped but were being
flagged as such.